requirements.txt
# AI Agents Intensive Course - Dependencies

# Google Agent Development Kit with A2A support
google-adk

# LiteLLM for third-party model integration (Doubao)
litellm

# Environment variable management
python-dotenv

# Async SQLite driver for DatabaseSessionService
aiosqlite

# Required for Day 5 A2A communication
requests
uvicorn
